rebase - add global-context index handler
